refactor: async FilePickerService
